Map action
1. Click action
2. Map color
3. Map -> Profile, Ticket
