+++ models/view/ViewModel.ts
@@ -3,5 +3,9 @@ import type { CodeModel } from "../data/ElectionModel";
 export type MapViewModel = CodeModel & {
     name: string[],
     path: string,
-    fill: string,
+    party_color: string,
+    party_code: string,
+    party_name: string
 }
+
+// name === party_name
